wip:  made populate file. Added DAO Tests. Working on logging and exception handling

+++ src/main/java/app/daos/UrlDAO.java
@@ -6,10 +6,7 @@ import app.entities.Url;
 import app.entities.UrlTracking;
 import app.security.entities.User;
 import app.utils.Conversion;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityManagerFactory;
-import jakarta.persistence.EntityNotFoundException;
-import jakarta.persistence.TypedQuery;
+import jakarta.persistence.*;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -39,48 +36,116 @@ public class UrlDAO implements IDao {
     public UrlDTO getLongUrl(String shortUrl) {
         Url url;
         try (EntityManager em = emf.createEntityManager()) {
-            TypedQuery<Url> query = em.createQuery("SELECT u FROM Url u WHERE u.shortUrl = :url", Url.class);
-            query.setParameter("url", shortUrl);
-            url = query.getSingleResult();
+//            TypedQuery<Url> query = em.createQuery("SELECT u FROM Url u WHERE u.shortUrl = :url", Url.class);
+//            query.setParameter("url", shortUrl);
+//            url = query.getSingleResult();
+            url = em.find(Url.class, shortUrl);
             if (url == null) {
-                throw new EntityNotFoundException();
+                throw new EntityNotFoundException(shortUrl);
             }
+
+        } catch (EntityNotFoundException e) {
+            System.err.println("ShortUrl Entity not found: " + e.getMessage());
+            throw e;
+        } catch (IllegalStateException e) {
+            System.err.println("An error occurred during the transaction: " + e.getMessage());
+            throw new IllegalStateException(e);
+        } catch (Exception e) {
+            System.err.println("An unexpected error occurred: " + e.getMessage());
+            throw new RuntimeException("Unexpected error occurred in getLongUrl method", e);
         }
         return new UrlDTO(url);
     }
 
     @Override
     public UrlDTO create(UrlDTO urlDTO, UserDTO userDTO) {
+        Url url;
         try (EntityManager em = emf.createEntityManager()) {
 
-            Url url = new Url(urlDTO);
-            User user = new User(userDTO);
-            url.addUser(user);
+            TypedQuery<User> query = em.createQuery("SELECT u FROM User u WHERE u.username = :name", User.class);
+            query.setParameter("name", userDTO.getUsername());
+            User user = query.getSingleResult();
 
+            url = new Url(urlDTO);
+            url.addUser(user);
             url.setShortUrl(Conversion.shortCode());
+
+            user.addUrl(url);
 
             em.getTransaction().begin();
             em.persist(url);
+            em.merge(user);
             em.getTransaction().commit();
-            return new UrlDTO(url);
 
+
+        } /*catch (NoResultException e) { //Not needed, because the user will never get here, because they are not logged in.
+            throw new EntityNotFoundException(e.getMessage());
+        }*/ catch (IllegalStateException e) {
+            System.err.println("An error occurred during the transaction: " + e.getMessage());
+            throw new IllegalStateException(e);
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            System.err.println("An unexpected error occurred: " + e.getMessage());
+            throw new RuntimeException("Unexpected error occurred in create method", e);
         }
+        return new UrlDTO(url);
     }
 
     @Override
     public void delete(String shortUrl) {
         try (EntityManager em = emf.createEntityManager()) {
+
             em.getTransaction().begin();
             Url url = em.find(Url.class, shortUrl);
+
+            if (url == null) {
+                throw new EntityNotFoundException(shortUrl);
+            }
+
             em.remove(url);
             em.getTransaction().commit();
+
+        } catch (EntityNotFoundException e) {
+            System.err.println("ShortUrl Entity not found: " + e.getMessage());
+            throw e;
+        } catch (IllegalStateException e) {
+            System.err.println("An error occurred during the transaction: " + e.getMessage());
+            throw e;
+        } catch (Exception e) {
+            System.err.println("An unexpected error occurred: " + e.getMessage());
+            throw new RuntimeException("Unexpected error occurred in delete method", e);
         }
     }
 
     @Override
-    public UrlDTO update() {
-        return null;
+    public UrlDTO updateLong(String shortUrl, UrlDTO urlDTO) {
+        Url urlFound;
+        try (EntityManager em = emf.createEntityManager()) {
+
+            urlFound = em.find(Url.class, shortUrl);
+
+            if (urlFound == null) {
+                throw new EntityNotFoundException(shortUrl);
+            }
+
+            urlFound.setLongUrl(urlDTO.getLongUrl());
+
+            em.getTransaction().begin();
+            em.merge(urlFound);
+            em.getTransaction().commit();
+
+        } catch (EntityNotFoundException e) {
+            System.err.println("Entity not found: " + e.getMessage());
+            throw e;
+        } catch (IllegalStateException e) {
+            System.err.println("An error occurred during the transaction: " + e.getMessage());
+            throw new IllegalStateException(e);
+        } catch (Exception e) {
+            System.err.println("An unexpected error occurred: " + e.getMessage());
+            throw new RuntimeException("Unexpected error occurred in updateLong method", e);
+        }
+        return new UrlDTO(urlFound);
     }
+
+
+
 }
